Struct deadpool_postgres::config::Config

pub struct Config {
    pub user: Option<String>,
    pub password: Option<String>,
    pub dbname: Option<String>,
    pub options: Option<String>,
    pub application_name: Option<String>,
    pub ssl_mode: Option<SslMode>,
    pub host: Option<String>,
    pub hosts: Option<Vec<String>>,
    pub port: Option<u16>,
    pub ports: Option<Vec<u16>>,
    pub connect_timeout: Option<Duration>,
    pub keepalives: Option<bool>,
    pub keepalives_idle: Option<Duration>,
    pub target_session_attrs: Option<TargetSessionAttrs>,
    pub channel_binding: Option<ChannelBinding>,
    pub manager: Option<ManagerConfig>,
    pub pool: Option<PoolConfig>,
}

Configuration object. By enabling the config feature you can read the configuration using the config crate.

Example environment

PG_HOST=pg.example.com
PG_USER=john_doe
PG_PASSWORD=topsecret
PG_DBNAME=example
PG_POOL.MAX_SIZE=16
PG_POOL.TIMEOUTS.WAIT.SECS=5
PG_POOL.TIMEOUTS.WAIT.NANOS=0

Example usage

Config::from_env("PG");

Fields
user: Option<String>

See tokio_postgres::Config::user

password: Option<String>

See tokio_postgres::Config::password

dbname: Option<String>

See tokio_postgres::Config::dbname

options: Option<String>

See tokio_postgres::Config::options

application_name: Option<String>

See tokio_postgres::Config::application_name

ssl_mode: Option<SslMode>

See tokio_postgres::Config::ssl_mode

host: Option<String>

This is similar to hosts but only allows one host to be specified. Unlike tokio-postgres::Config this structure differenciates between one host and more than one host. This makes it possible to store this configuration in an envorinment variable. See tokio_postgres::Config::host

hosts: Option<Vec<String>>

See tokio_postgres::Config::hosts

port: Option<u16>

This is similar to ports but only allows one port to be specified. Unlike tokio-postgres::Config this structure differenciates between one port and more than one port. This makes it possible to store this configuration in an environment variable. See tokio_postgres::Config::port

ports: Option<Vec<u16>>

See tokio_postgres::Config::port

connect_timeout: Option<Duration>

See tokio_postgres::Config::connect_timeout

keepalives: Option<bool>

See tokio_postgres::Config::keepalives

keepalives_idle: Option<Duration>

See tokio_postgres::Config::keepalives_idle

target_session_attrs: Option<TargetSessionAttrs>

See tokio_postgres::Config::target_session_attrs

channel_binding: Option<ChannelBinding>

See tokio_postgres::Config::channel_binding

manager: Option<ManagerConfig>

Manager configuration

pool: Option<PoolConfig>

Pool configuration

Methods

impl Config

pub fn from_env(prefix: &str) -> Result<Self, ConfigError>

Deprecated since 0.5.5:
Please embed this structure in your own config structure and use config::Config directly.

Create configuration from environment variables.

pub fn create_pool<T>(&self, tls: T) -> Result<Pool, ConfigError> where
T: MakeTlsConnect<Socket> + Clone + Sync + Send + 'static,
T::Stream: Sync + Send,
T::TlsConnect: Sync + Send,
<T::TlsConnect as TlsConnect<Socket>>::Future: Send,

Create pool using the current configuration

pub fn get_pg_config(&self) -> Result<Config, ConfigError>

Get tokio_postgres::Config which can be used to connect to the database.

pub fn get_manager_config(&self) -> ManagerConfig

Get deadpool_postgres::ManagerConfig which can be used to construct a deadpool::managed::Pool instance.

pub fn get_pool_config(&self) -> PoolConfig

Get deadpool::PoolConfig which can be used to construct a deadpool::managed::Pool instance.